Crate fedimint_client

Client library

Notably previous crate became fedimint_client_module and the project is gradually moving things, that are irrelevant to the interface between client and client modules.

Modules

api

Re-exporting of everything from fedimint_client_module

api_announcements

Re-exporting of everything from fedimint_client_module

backup

Re-exporting of everything from fedimint_client_module

db

Re-exporting of everything from fedimint_client_module

derivable_secret

Re-exporting of everything from fedimint_client_module

envs

Re-exporting of everything from fedimint_client_module

meta

Re-exporting of everything from fedimint_client_module

module

Re-exporting of everything from fedimint_client_module

oplog

Re-exporting of everything from fedimint_client_module

secret

Re-exporting of everything from fedimint_client_module

sm

Re-exporting of everything from fedimint_client_module

transaction

Re-exporting of everything from fedimint_client_module

Macros

sm_enum_variant_translation

Re-exporting of everything from fedimint_client_module

Structs

AdminCreds

Re-exporting of everything from fedimint_client_module

Client

Re-exporting of everything from fedimint_client_module

ClientBuilder

Re-exporting of everything from fedimint_client_module

ClientHandle

Re-exporting of everything from fedimint_client_module

ClientModuleInstance

Re-exporting of everything from fedimint_client_module

DynGlobalClientContext

Re-exporting of everything from fedimint_client_module

ModuleRecoveryCompleted

Re-exporting of everything from fedimint_client_module

ModuleRecoveryStarted

Re-exporting of everything from fedimint_client_module

TransactionUpdates

Re-exporting of everything from fedimint_client_module

TxAcceptedEvent

Re-exporting of everything from fedimint_client_module

TxCreatedEvent

Re-exporting of everything from fedimint_client_module

TxRejectedEvent

Re-exporting of everything from fedimint_client_module

Enums

AddStateMachinesError

Re-exporting of everything from fedimint_client_module

Traits

IGlobalClientContext

Re-exporting of everything from fedimint_client_module

Functions

client_decoders

Re-exporting of everything from fedimint_client_module

get_decoded_client_secret

Re-exporting of everything from fedimint_client_module

Type Aliases

AddStateMachinesResult

Re-exporting of everything from fedimint_client_module

ClientHandleArc

Re-exporting of everything from fedimint_client_module

InstancelessDynClientInput

Re-exporting of everything from fedimint_client_module

InstancelessDynClientInputBundle

Re-exporting of everything from fedimint_client_module

InstancelessDynClientInputSM

Re-exporting of everything from fedimint_client_module

InstancelessDynClientOutput

Re-exporting of everything from fedimint_client_module

InstancelessDynClientOutputBundle

Re-exporting of everything from fedimint_client_module

InstancelessDynClientOutputSM

Re-exporting of everything from fedimint_client_module

ModuleGlobalContextGen

Re-exporting of everything from fedimint_client_module