fedimint_client

Module api_announcements

Federation Api announcement handling

Structs

ApiAnnouncementKey

ApiAnnouncementPrefix

Functions

fetch_api_announcements_from_at_least_num_of_peers 🔒

Fetch responses from at least num_responses_required of peers.

get_api_urls

Returns a list of all peers and their respective API URLs taking into account announcements overwriting the URLs contained in the original configuration.

run_api_announcement_refresh_task 🔒

Fetches API URL announcements from guardians, validates them and updates the DB if any new more upt to date ones are found.

store_api_announcement_updates 🔒

store_api_announcements_updates_from_peers 🔒

Type Aliases

PeersSignedApiAnnouncements 🔒