fedimint_client

Module client

Core Client

Modules

builder 🔒

event_log 🔒

global_ctx 🔒

handle 🔒

Structs

Client

Main client type

GetBalanceChangesRequest

GetOperationIdRequest

ListOperationsParams 🔒

PrimaryModuleCandidates 🔒

Primary module candidates at specific priority level

Constants

SUPPORTED_CORE_API_VERSIONS 🔒

List of core api versions supported by the implementation. Notably major version is the one being supported, and corresponding minor version is the one required (for given major version).

Functions

client_decoders