fedimint_client::client::handle

Struct ClientHandle

pub struct ClientHandle {
    inner: Option<Arc<Client>>,
}

User handle to the Client instance

On the drop of ClientHandle the client will be shut-down, and resources it used freed.

Notably it ops::Derefs to the Client where most methods live.

Put this in an Arc to clone it (see ClientHandleArc).

Fields
inner: Option<Arc<Client>>
Implementations

impl ClientHandle

pub(crate) fn new(inner: Arc<Client>) -> Self

Create

pub(crate) fn as_inner(&self) -> &Arc<Client>

pub fn start_executor(&self)

pub async fn shutdown(self)

Shutdown the client.

async fn shutdown_inner(&mut self)

pub async fn restart(self) -> Result<ClientHandle>

Restart the client

Returns false if there are other clones of ClientHandle, or starting the client again failed for some reason.

Notably it will re-use the original fedimint_core::db::Database handle, and not attempt to open it again.