fedimint_client
Function get_decoded_client_secret

pub async fn get_decoded_client_secret<T: Decodable>(db: &Database) -> Result<T>

Fetches the encoded client secret from the database and decodes it. If an encoded client secret is not present in the database, or if decoding fails, an error is returned.