fedimint_client

Module guardian_metadata

Guardian metadata handling

Structs

GuardianMetadataKey

GuardianMetadataPrefix

Functions

fetch_guardian_metadata_from_at_least_num_of_peers 🔒

Fetch responses from at least num_responses_required of peers.

run_guardian_metadata_refresh_task 🔒

Fetches guardian metadata from guardians, validates them and updates the DB if any new more up to date ones are found.

store_guardian_metadata_updates 🔒

store_guardian_metadata_updates_from_peers 🔒

Type Aliases

PeersSignedGuardianMetadata 🔒