Module fedimint_client::module

Module client interface definitions

Modules

init

recovery

Structs

ClientContext

A Client context for a ClientModule M

ClientContextSelfRef

A reference back to itself that the module cacn get from the ClientContext

ClientDbTxContext

A context of a database transaction started with ClientContext::module_autocommit

DynClientModule

FinalClient

A final, fully initialized crate::Client

Traits

ClientModule

Fedimint module client

IClientModule

Type-erased version of ClientModule

Type Aliases

ClientModuleRegistry

StateGenerator