Trait fedimint_client::module::gen::ClientModuleGen

pub trait ClientModuleGen: ExtendsCommonModuleGen + Sized {
    type Module: ClientModule;
    type Config: TypedClientModuleConfig;

    fn init<'life0, 'async_trait>(
        &'life0 self,
        cfg: Self::Config,
        db: Database,
        instance_id: ModuleInstanceId
    ) -> Pin<Box<dyn Future<Output = Result<Self::Module>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn init_primary<'life0, 'async_trait>(
        &'life0 self,
        _cfg: Self::Config,
        _db: Database
    ) -> Pin<Box<dyn Future<Output = Result<DynPrimaryClientModule>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
}

Required Associated Types

type Module: ClientModule

type Config: TypedClientModuleConfig

Required Methods

fn init<'life0, 'async_trait>(
&'life0 self,
cfg: Self::Config,
db: Database,
instance_id: ModuleInstanceId
) -> Pin<Box<dyn Future<Output = Result<Self::Module>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Initialize a ClientModule instance from its config

Provided Methods

fn init_primary<'life0, 'async_trait>(
&'life0 self,
_cfg: Self::Config,
_db: Database
) -> Pin<Box<dyn Future<Output = Result<DynPrimaryClientModule>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Initialize a crate::module::PrimaryClientModule instance from its config

The default implementation returns an error, assuming that the module is not a primary one. If it is the default impl has to be overridden as follows:

async fn init_primary(
    &self,
    cfg: Self::Config,
    db: Database,
) -> anyhow::Result<DynPrimaryClientModule> {
    Ok(self.init(cfg, db)?.into())
}