Trait fedimint_client::module::PrimaryClientModule

pub trait PrimaryClientModule: ClientModule {
    fn create_sufficient_input<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        min_amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<(Vec<KeyPair>, <Self::Common as ModuleCommon>::Input, StateGenerator<Self::States>)>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn create_exact_output<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        amount: Amount
    ) -> Pin<Box<dyn Future<Output = (<Self::Common as ModuleCommon>::Output, StateGenerator<Self::States>)> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
}

A client module that can be used as funding source and to generate arbitrary change outputs for unbalanced transactions.

Required Methods

fn create_sufficient_input<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut DatabaseTransaction<'life2>,
min_amount: Amount
) -> Pin<Box<dyn Future<Output = Result<(Vec<KeyPair>, <Self::Common as ModuleCommon>::Input, StateGenerator<Self::States>)>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Creates an input of at least a given min_amount from the holdings managed by the module.

If successful it returns:

A set of private keys belonging to the input for signing the transaction

The input of at least min_amount, the actual amount might be larger, the caller has to handle this case and possibly generate change using create_change_output.

A closure that generates states belonging to the input. This closure takes the transaction id of the transaction in which the input was used and the input index as input since these cannot be known at time of calling create_funding_input and have to be injected later.

The function returns an error if the client’s funds are not sufficient to create the requested input.

fn create_exact_output<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut DatabaseTransaction<'life2>,
amount: Amount
) -> Pin<Box<dyn Future<Output = (<Self::Common as ModuleCommon>::Output, StateGenerator<Self::States>)> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Creates an output of exactly amount that will pay into the holdings managed by the module.

It returns:

The output of exactly amount.

A closure that generates states belonging to the output. This closure takes the transaction id of the transaction in which the output was used and the output index as input since these cannot be known at time of calling create_change_output and have to be injected later.