fedimint_client
Module secret

Secret handling & derivation

Structs

PlainRootSecretStrategy

Just uses 64 random bytes and derives the secret from them

Constants

TYPE_BACKUP 🔒

TYPE_MODULE 🔒

TYPE_PRE_ROOT_SECRET_HASH 🔒

Traits

DeriveableSecretClientExt

RootSecretStrategy

Trait defining a way to generate, serialize and deserialize a root secret. It defines a Encoding associated type which represents a specific representation of a secret (e.g. a bip39, slip39, CODEX32, … struct) and then defines the methods necessary for the client to interact with it.

Functions

get_default_client_secret

Convenience function to derive fedimint-client root secret using the default (0) wallet number, given a global root secret that’s managed externally by a consumer of fedimint-client.