fedimint_client::sm

Module executor

Structs

ActiveModuleOperationStateKeyPrefix 🔒

ActiveOperationStateKeyPrefix

ActiveStateKeyBytes

ActiveStateKeyDb

ActiveStateKeyPrefix

ActiveStateKeyPrefixBytes

Executor

Executor that drives forward state machines under its management.

ExecutorBuilder

Builder to which module clients can be attached and used to build an Executor supporting these.

ExecutorInner 🔒

InactiveModuleOperationStateKeyPrefix 🔒

InactiveOperationStateKeyPrefix

InactiveStateKeyBytes

InactiveStateKeyDb

InactiveStateKeyPrefix

InactiveStateKeyPrefixBytes

StateMachineUpdated

TransitionForActiveState 🔒

Enums

ActiveOrInactiveState 🔒

ExecutorDbPrefixes 🔒

Prefixes for executor DB entries

ExecutorState 🔒

Constants

MAX_DB_ATTEMPTS 🔒

After how many attempts a DB transaction is aborted with an error