fedimint_client::sm
Module state

State machine state interface

Structs

DynContext

A shared context for a module client state machine

DynState

A type-erased state of a state machine belonging to a module instance, see State

OperationState

StateTransition

Represents one or multiple possible state transitions triggered in a common way

Traits

Context

Additional data made available to state machines of a module (e.g. API clients)

IContext

Something that can be a DynContext for a state machine

IState

Object-safe version of State

State

Implementors act as state machines that can be executed

Type Aliases

StateTransitionFunction 🔒

TriggerFuture 🔒