Trait fedimint_client::sm::State

pub trait State: Debug + Clone + Eq + PartialEq + Encodable + Decodable + MaybeSend + MaybeSync + 'static {
    type ModuleContext: Context;
    type GlobalContext: GlobalContext;

    fn transitions(
        &self,
        context: &Self::ModuleContext,
        global_context: &Self::GlobalContext
    ) -> Vec<StateTransition<Self>> ;
    fn operation_id(&self) -> OperationId;
}

Implementors act as state machines that can be executed

Required Associated Types

type ModuleContext: Context

Additional resources made available in this module’s state transitions

type GlobalContext: GlobalContext

Additional resources made available for all state transitions

Required Methods

fn transitions(
&self,
context: &Self::ModuleContext,
global_context: &Self::GlobalContext
) -> Vec<StateTransition<Self>>

All possible transitions from the current state to other states. See StateTransition for details.

fn operation_id(&self) -> OperationId

Operation this state machine belongs to. See OperationId for details.