fedimint_client

Struct ClientBuilder

pub struct ClientBuilder {
    module_inits: ClientModuleInitRegistry,
    admin_creds: Option<AdminCreds>,
    meta_service: Arc<MetaService>,
    stopped: bool,
    log_event_added_transient_tx: Sender<EventLogEntry>,
    request_hook: ApiRequestHook,
    iroh_enable_dht: bool,
    iroh_enable_next: bool,
}

Used to configure, assemble and build Client

Fields
module_inits: ClientModuleInitRegistryadmin_creds: Option<AdminCreds>meta_service: Arc<MetaService>stopped: boollog_event_added_transient_tx: Sender<EventLogEntry>request_hook: ApiRequestHookiroh_enable_dht: booliroh_enable_next: bool
Implementations

impl ClientBuilder

pub(crate) fn new() -> Self

pub(crate) fn from_existing(client: &Client) -> Self

pub fn with_module_inits(&mut self, module_inits: ClientModuleInitRegistry)

Replace module generator registry entirely

pub fn with_module<M: ClientModuleInit>(&mut self, module_init: M)

Make module generator available when reading the config

pub fn stopped(&mut self)

pub fn with_api_request_hook(self, hook: ApiRequestHook) -> Self

Build the Client with a custom wrapper around its api request logic

This is intended to be used by downstream applications, e.g. to:

simulate offline mode,

save battery when the OS indicates lack of connectivity,

inject faults and delays for testing purposes,

collect statistics and emit notifications.

pub fn with_meta_service(&mut self, meta_service: Arc<MetaService>)

pub fn with_iroh_enable_dht(self, iroh_enable_dht: bool) -> Self

Override if the DHT should be enabled when using Iroh to connect to the federation

pub fn with_iroh_enable_next(self, iroh_enable_next: bool) -> Self

Override if the parallel unstable/next Iroh stack should be enabled when using Iroh to connect to the federation

async fn migrate_module_dbs( &self, db: &Database, client_config: &ClientConfig, ) -> Result<()>

Migrate client module databases

Note: Client core db migration are done immediately in Client::builder, to ensure db matches the code at all times, while migrating modules requires figuring out what modules actually are first.

pub async fn load_existing_config(&self, db: &Database) -> Result<ClientConfig>

pub fn set_admin_creds(&mut self, creds: AdminCreds)

async fn init( self, connectors: ConnectorRegistry, db_no_decoders: Database, pre_root_secret: DerivableSecret, config: ClientConfig, api_secret: Option<String>, init_mode: InitMode, preview_prefetch_api_announcements: Option<Jit<Vec<BTreeMap<PeerId, SignedApiAnnouncement>>>>, preview_prefetch_api_version_set: Option<JitTryAnyhow<BTreeMap<PeerId, SupportedApiVersionsSummary>>>, prefetch_chain_id: Option<JitTryAnyhow<ChainId>>, ) -> Result<ClientHandle>

pub async fn preview( self, connectors: ConnectorRegistry, invite_code: &InviteCode, ) -> Result<ClientPreview>

pub async fn preview_with_existing_config( self, connectors: ConnectorRegistry, config: ClientConfig, api_secret: Option<String>, ) -> Result<ClientPreview>

Use Self::preview instead

If reuse_api is set, it will allow the preview to prefetch some data to speed up the final join.

async fn preview_inner( self, connectors: ConnectorRegistry, config: ClientConfig, api_secret: Option<String>, prefetch_api: Option<DynGlobalApi>, prefetch_api_announcements: Option<Jit<Vec<BTreeMap<PeerId, SignedApiAnnouncement>>>>, ) -> Result<ClientPreview>

pub async fn open( self, connectors: ConnectorRegistry, db_no_decoders: Database, pre_root_secret: RootSecret, ) -> Result<ClientHandle>

pub(crate) async fn build( self, connectors: ConnectorRegistry, db_no_decoders: Database, pre_root_secret: DerivableSecret, config: ClientConfig, api_secret: Option<String>, stopped: bool, preview_prefetch_api_announcements: Option<Jit<Vec<BTreeMap<PeerId, SignedApiAnnouncement>>>>, preview_prefetch_api_version_set: Option<JitTryAnyhow<BTreeMap<PeerId, SupportedApiVersionsSummary>>>, prefetch_chain_id: Option<JitTryAnyhow<ChainId>>, ) -> Result<ClientHandle>

Build a Client and start the executor

async fn build_stopped( self, connectors: ConnectorRegistry, db_no_decoders: Database, pre_root_secret: DerivableSecret, config: &ClientConfig, api_secret: Option<String>, log_event_added_transient_tx: Sender<EventLogEntry>, request_hook: ApiRequestHook, preview_prefetch_api_announcements: Option<Jit<Vec<BTreeMap<PeerId, SignedApiAnnouncement>>>>, preview_prefetch_api_version_set: Option<JitTryAnyhow<BTreeMap<PeerId, SupportedApiVersionsSummary>>>, prefetch_chain_id: Option<JitTryAnyhow<ChainId>>, ) -> Result<ClientHandle>

Build a Client but do not start the executor

async fn load_init_state(db: &Database) -> InitState

fn decoders(&self, config: &ClientConfig) -> ModuleDecoderRegistry

fn config_decoded( config: &ClientConfig, decoders: &ModuleDecoderRegistry, ) -> Result<ClientConfig, DecodeError>

fn federation_root_secret( pre_root_secret: &DerivableSecret, config: &ClientConfig, ) -> DerivableSecret

Re-derive client’s root_secret using the federation ID. This eliminates the possibility of having the same client root_secret across multiple federations.

pub fn get_event_log_transient_receiver(&self) -> Receiver<EventLogEntry>

Register to receiver all new transient (unpersisted) events

async fn migrate_pending_config_if_present(db: &Database)

Check for pending config and migrate it if present. Returns the config to use (either the original or the migrated pending config).

fn load_and_refresh_client_config_static( config: &ClientConfig, api: &DynGlobalApi, db: &Database, task_group: &TaskGroup, )

Asynchronously refetch client config from federation and compare with existing. If different, save to pending config in database.

async fn refresh_client_config_static( config: &ClientConfig, api: &DynGlobalApi, db: &Database, )

Wrapper that handles errors from config refresh with proper logging

fn validate_config_update( current_config: &ClientConfig, new_config: &ClientConfig, ) -> Result<()>

Validate that a config update is valid

async fn refresh_client_config_static_try( current_config: &ClientConfig, api: &DynGlobalApi, db: &Database, ) -> Result<()>

Refetch client config from federation and save as pending if different