Struct fedimint_client::ClientInner

pub(crate) struct ClientInner {
    pub(crate) config: ClientConfig,
    pub(crate) decoders: ModuleDecoderRegistry,
    pub(crate) db: Database,
    pub(crate) federation_id: FederationId,
    pub(crate) federation_meta: BTreeMap<String, String>,
    pub(crate) primary_module_instance: ModuleInstanceId,
    pub(crate) modules: ClientModuleRegistry,
    pub(crate) executor: Executor<DynGlobalClientContext>,
    pub(crate) api: DynGlobalApi,
    pub(crate) root_secret: DerivableSecret,
    pub(crate) operation_log: OperationLog,
    pub(crate) secp_ctx: Secp256k1<All>,
}

Fields
config: ClientConfigdecoders: ModuleDecoderRegistrydb: Databasefederation_id: FederationIdfederation_meta: BTreeMap<String, String>primary_module_instance: ModuleInstanceIdmodules: ClientModuleRegistryexecutor: Executor<DynGlobalClientContext>api: DynGlobalApiroot_secret: DerivableSecretoperation_log: OperationLogsecp_ctx: Secp256k1<All>
Implementations

impl ClientInner

pub(crate) fn primary_module(&self) -> &DynClientModule

pub(crate) fn context_gen(self: &Arc<Self>) -> ModuleGlobalContextGen

pub(crate) fn config(&self) -> &ClientConfig

pub(crate) fn decoders(&self) -> &ModuleDecoderRegistry

pub(crate) fn get_module(
&self,
instance: ModuleInstanceId
) -> &(dyn IClientModule + Send + Sync)

Returns a reference to the module, panics if not found

pub(crate) fn try_get_module(
&self,
instance: ModuleInstanceId
) -> Option<&(dyn IClientModule + Send + Sync)>

pub(crate) fn transaction_builder_balance(
&self,
builder: &TransactionBuilder
) -> TransactionBuilderBalance

Determines if a transaction is underfunded, overfunded or balanced

pub(crate) async fn finalize_transaction(
&self,
dbtx: &mut DatabaseTransaction<'_>,
operation_id: OperationId,
partial_transaction: TransactionBuilder
) -> Result<(Transaction, Vec<DynState<DynGlobalClientContext>>, Option<u64>)>

Adds funding to a transaction or removes overfunding via change.

pub(crate) async fn finalize_and_submit_transaction(
&self,
dbtx: &mut DatabaseTransaction<'_>,
operation_id: OperationId,
tx_builder: TransactionBuilder
) -> Result<(TransactionId, Option<OutPoint>)>

pub(crate) async fn transaction_update_stream(
&self,
operation_id: OperationId
) -> BoxStream<'static, OperationState<TxSubmissionStates>>

pub(crate) async fn operation_exists(
dbtx: &mut DatabaseTransaction<'_>,
operation_id: OperationId
) -> bool

pub async fn await_primary_module_output(
&self,
operation_id: OperationId,
out_point: OutPoint
) -> Result<Amount>