Module fedimint_client::db

Database keys used by the client

Structs

CachedApiVersionSet

CachedApiVersionSetKey

ChronologicalOperationLogKey

Key used to lookup operation log entries in chronological order

ChronologicalOperationLogKeyPrefix

ClientConfigKey

ClientConfigKeyPrefix

ClientInitStateKey

Does the client modules need to run recovery before being usable?

ClientInitStatePrefix

ClientInviteCodeKey

ClientInviteCodeKeyPrefix

ClientMetadataKey

Client metadata that will be stored/restored on backup&recovery

ClientMetadataPrefix

ClientModuleRecovery

ClientModuleRecoveryPrefix

ClientModuleRecoveryState

ClientRecoverySnapshot

ClientRecoverySnapshotPrefix

DbKeyPrefixIter

An iterator over the variants of Self

EncodedClientSecretKey

EncodedClientSecretKeyPrefix

OperationLogKey

Enums

DbKeyPrefix

InitMode

Client initialization mode

InitModeComplete

Like InitMode, but without no longer required data.

InitState

The state of the client initialization