Module fedimint_client::meta

Management of meta fields

Structs

LegacyMetaSource

Legacy non-meta module config source uses client config meta and meta_override_url meta field.

MetaService

Service for managing the caching of meta fields.

MetaValue

MetaValues

Enums

FetchKind

Traits

MetaSource

Functions

fetch_meta_overrides

parse_meta_value_static

Tries to parse str_value as JSON. In the special case that V is String we return the raw str_value if JSON parsing fails. This necessary since the spec wasn’t clear enough in the beginning.