Trait fedimint_client::module::ClientModule

pub trait ClientModule: Debug + MaybeSend + MaybeSync + 'static {
    type Common: ModuleCommon;
    type ModuleStateMachineContext: Context;
    type GlobalStateMachineContext: GlobalContext;
    type States: State<Self::GlobalStateMachineContext, ModuleContext = Self::ModuleStateMachineContext>;

    fn context(&self) -> Self::ModuleStateMachineContext;

    fn decoder() -> Decoder { ... }
}

Fedimint module client

Required Associated Types

type Common: ModuleCommon

Common module types shared between client and server

type ModuleStateMachineContext: Context

Data and API clients available to state machine transitions of this module

type GlobalStateMachineContext: GlobalContext

Data and API clients available to state machine transitions of all modules

type States: State<Self::GlobalStateMachineContext, ModuleContext = Self::ModuleStateMachineContext>

All possible states this client can submit to the executor

Required Methods

fn context(&self) -> Self::ModuleStateMachineContext

Provided Methods

fn decoder() -> Decoder