Trait fedimint_client::module::ClientModule

pub trait ClientModule: Debug + MaybeSend + MaybeSync + 'static {
    type Common: ModuleCommon;
    type ModuleStateMachineContext: Context;
    type States: State<GlobalContext = DynGlobalClientContext, ModuleContext = Self::ModuleStateMachineContext> + IntoDynInstance<DynType = DynState<DynGlobalClientContext>>;

    fn context(&self) -> Self::ModuleStateMachineContext;
    fn input_amount(
        &self,
        input: &<Self::Common as ModuleCommon>::Input
    ) -> TransactionItemAmount;
    fn output_amount(
        &self,
        output: &<Self::Common as ModuleCommon>::Output
    ) -> TransactionItemAmount;

    fn decoder() -> Decoder { ... }
    fn handle_cli_command<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _client: &'life1 Client,
        _args: &'life2 [OsString]
    ) -> Pin<Box<dyn Future<Output = Result<Value>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
    { ... }
    fn supports_backup(&self) -> bool { ... }
    fn backup<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
        _executor: Executor<DynGlobalClientContext>,
        _api: DynGlobalApi,
        _module_instance_id: ModuleInstanceId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
    { ... }
    fn restore<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransaction<'life2>,
        _module_instance_id: ModuleInstanceId,
        _executor: Executor<DynGlobalClientContext>,
        _api: DynGlobalApi,
        _snapshot: Option<&'life3 [u8]>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
    { ... }
    fn wipe<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
        _module_instance_id: ModuleInstanceId,
        _executor: Executor<DynGlobalClientContext>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
    { ... }
}

Fedimint module client

Required Associated Types

type Common: ModuleCommon

Common module types shared between client and server

type ModuleStateMachineContext: Context

Data and API clients available to state machine transitions of this module

type States: State<GlobalContext = DynGlobalClientContext, ModuleContext = Self::ModuleStateMachineContext> + IntoDynInstance<DynType = DynState<DynGlobalClientContext>>

All possible states this client can submit to the executor

Required Methods

fn context(&self) -> Self::ModuleStateMachineContext

fn input_amount(
&self,
input: &<Self::Common as ModuleCommon>::Input
) -> TransactionItemAmount

Returns the amount represented by the input and the fee its processing requires

fn output_amount(
&self,
output: &<Self::Common as ModuleCommon>::Output
) -> TransactionItemAmount

Returns the amount represented by the output and the fee its processing requires

Provided Methods

fn decoder() -> Decoder

fn handle_cli_command<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
_client: &'life1 Client,
_args: &'life2 [OsString]
) -> Pin<Box<dyn Future<Output = Result<Value>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

fn supports_backup(&self) -> bool

fn backup<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
_dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
_executor: Executor<DynGlobalClientContext>,
_api: DynGlobalApi,
_module_instance_id: ModuleInstanceId
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

fn restore<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
_dbtx: &'life1 mut DatabaseTransaction<'life2>,
_module_instance_id: ModuleInstanceId,
_executor: Executor<DynGlobalClientContext>,
_api: DynGlobalApi,
_snapshot: Option<&'life3 [u8]>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

fn wipe<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
_dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
_module_instance_id: ModuleInstanceId,
_executor: Executor<DynGlobalClientContext>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,