Trait fedimint_client::module::ClientModule

pub trait ClientModule: Debug + MaybeSend + MaybeSync + 'static {
    type Init: ClientModuleInit;
    type Common: ModuleCommon;
    type ModuleStateMachineContext: Context;
    type States: State<GlobalContext = DynGlobalClientContext, ModuleContext = Self::ModuleStateMachineContext> + IntoDynInstance<DynType = DynState<DynGlobalClientContext>>;

    // Required methods
    fn context(&self) -> Self::ModuleStateMachineContext;
    fn input_amount(
        &self,
        input: &<Self::Common as ModuleCommon>::Input
    ) -> TransactionItemAmount;
    fn output_amount(
        &self,
        output: &<Self::Common as ModuleCommon>::Output
    ) -> TransactionItemAmount;

    // Provided methods
    fn decoder() -> Decoder { ... }
    fn kind() -> ModuleKind { ... }
    fn handle_cli_command<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _client: &'life1 ClientArc,
        _args: &'life2 [OsString]
    ) -> Pin<Box<dyn Future<Output = Result<Value>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn supports_backup(&self) -> bool { ... }
    fn backup<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>,
        _executor: Executor<DynGlobalClientContext>,
        _api: DynGlobalApi,
        _module_instance_id: ModuleInstanceId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn restore<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransaction<'life2>,
        _module_instance_id: ModuleInstanceId,
        _executor: Executor<DynGlobalClientContext>,
        _api: DynGlobalApi,
        _snapshot: Option<&'life3 [u8]>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn wipe<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>,
        _module_instance_id: ModuleInstanceId,
        _executor: Executor<DynGlobalClientContext>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn supports_being_primary(&self) -> bool { ... }
    fn create_sufficient_input<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>,
        _operation_id: OperationId,
        _min_amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ClientInput<<Self::Common as ModuleCommon>::Input, Self::States>>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn create_exact_output<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>,
        _operation_id: OperationId,
        _amount: Amount
    ) -> Pin<Box<dyn Future<Output = Vec<ClientOutput<<Self::Common as ModuleCommon>::Output, Self::States>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn await_primary_module_output<'life0, 'async_trait>(
        &'life0 self,
        _operation_id: OperationId,
        _out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Result<Amount>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_balance<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>
    ) -> Pin<Box<dyn Future<Output = Amount> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn subscribe_balance_changes<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = BoxStream<'static, ()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn leave<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _dbtx: &'life1 mut DatabaseTransactionRef<'life2>,
        _module_instance_id: ModuleInstanceId,
        _executor: Executor<DynGlobalClientContext>,
        _api: DynGlobalApi
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }

}

Fedimint module client

Required Associated Types

type Init: ClientModuleInit

type Common: ModuleCommon

Common module types shared between client and server

type ModuleStateMachineContext: Context

Data and API clients available to state machine transitions of this module

type States: State<GlobalContext = DynGlobalClientContext, ModuleContext = Self::ModuleStateMachineContext> + IntoDynInstance<DynType = DynState<DynGlobalClientContext>>

All possible states this client can submit to the executor

Required Methods

fn context(&self) -> Self::ModuleStateMachineContext

fn input_amount( &self, input: &<Self::Common as ModuleCommon>::Input ) -> TransactionItemAmount

Returns the amount represented by the input and the fee its processing requires

fn output_amount( &self, output: &<Self::Common as ModuleCommon>::Output ) -> TransactionItemAmount

Returns the amount represented by the output and the fee its processing requires

Provided Methods

fn decoder() -> Decoder

fn kind() -> ModuleKind

fn handle_cli_command<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _client: &'life1 ClientArc, _args: &'life2 [OsString] ) -> Pin<Box<dyn Future<Output = Result<Value>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn supports_backup(&self) -> bool

fn backup<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2>, _executor: Executor<DynGlobalClientContext>, _api: DynGlobalApi, _module_instance_id: ModuleInstanceId ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn restore<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransaction<'life2>, _module_instance_id: ModuleInstanceId, _executor: Executor<DynGlobalClientContext>, _api: DynGlobalApi, _snapshot: Option<&'life3 [u8]> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

fn wipe<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2>, _module_instance_id: ModuleInstanceId, _executor: Executor<DynGlobalClientContext> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn supports_being_primary(&self) -> bool

Does this module support being a primary module

If it does it must implement:

Self::create_sufficient_input

Self::create_exact_output

Self::await_primary_module_output

Self::get_balance

Self::subscribe_balance_changes

fn create_sufficient_input<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2>, _operation_id: OperationId, _min_amount: Amount ) -> Pin<Box<dyn Future<Output = Result<Vec<ClientInput<<Self::Common as ModuleCommon>::Input, Self::States>>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Creates an input of at least a given min_amount from the holdings managed by the module.

If successful it returns:

A set of private keys belonging to the input for signing the transaction

The input of at least min_amount, the actual amount might be larger, the caller has to handle this case and possibly generate change using create_change_output.

A closure that generates states belonging to the input. This closure takes the transaction id of the transaction in which the input was used and the input index as input since these cannot be known at time of calling create_funding_input and have to be injected later.

The function returns an error if the client’s funds are not sufficient to create the requested input.

fn create_exact_output<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2>, _operation_id: OperationId, _amount: Amount ) -> Pin<Box<dyn Future<Output = Vec<ClientOutput<<Self::Common as ModuleCommon>::Output, Self::States>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Creates an output of exactly amount that will pay into the holdings managed by the module.

It returns:

The output of exactly amount.

A closure that generates states belonging to the output. This closure takes the transaction id of the transaction in which the output was used and the output index as input since these cannot be known at time of calling create_change_output and have to be injected later.

fn await_primary_module_output<'life0, 'async_trait>( &'life0 self, _operation_id: OperationId, _out_point: OutPoint ) -> Pin<Box<dyn Future<Output = Result<Amount>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Waits for the funds from an output created by Self::create_exact_output to become available. This function returning typically implies a change in the output of Self::get_balance.

fn get_balance<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2> ) -> Pin<Box<dyn Future<Output = Amount> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Returns the balance held by this module and available for funding transactions.

fn subscribe_balance_changes<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = BoxStream<'static, ()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns a stream that will output the updated module balance each time it changes.

fn leave<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _dbtx: &'life1 mut DatabaseTransactionRef<'life2>, _module_instance_id: ModuleInstanceId, _executor: Executor<DynGlobalClientContext>, _api: DynGlobalApi ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Leave the federation

While technically there’s nothing stopping the client from just abandoning Federation at any point by deleting all the related local data, it is useful to make sure it’s safe beforehand.

This call indicates the desire of the caller client code to orderly and safely leave the Federation by this module instance. The goal of the implementations is to fulfil that wish, giving prompt and informative feedback if it’s not yet possible.

The client module implementation should handle the request and return as fast as possible avoiding blocking for longer than necessary. This would usually involve some combination of:

recording the state of being in process of leaving the Federation to prevent initiating new conditions that could delay its completion;

performing any fast to complete cleanup/exit logic;

initiating any time-consuming logic (e.g. canceling outstanding contracts), as background jobs, tasks machines, etc.

checking for any conditions indicating it might not be safe to leave at the moment.

This function should return Ok only if from the perspective of this module instance, it is safe to delete client data and stop using it, with no further actions (like background jobs) required to complete.

This function should return an error if it’s not currently possible to safely (e.g. without loosing funds) leave the Federation. It should avoid running indefinitely trying to complete any cleanup actions necessary to reach a clean state, preferring spawning new state machines and returning an informative error about cleanup still in progress.

If any internal task needs to complete, any user action is required, or even external condition needs to be met this function should return a Err.

Notably modules should not disable interaction that might be necessary for the user (possibly through other modules) to leave the Federation. In particular a Mint module should retain ability to create new notes, and LN module should retain ability to send funds out.

Calling code must NOT assume that a module that once returned Ok, will not return Err at later point. E.g. a Mint module might have no outstanding balance at first, but other modules winding down might “cash-out” to Ecash.

Before leaving the Federation and deleting any state the calling code must collect a full round of Ok from all the modules.

Calling code should allow the user to override and ignore any outstanding errors, after sufficient amount of warnings. Ideally, this should be done on per-module basis, to avoid mistakes.