Module fedimint_client::sm

Client state machine interfaces and executor implementation

Modules

executor 🔒

state 🔒

State machine state interface

Structs

DynContext

A shared context for a module client state machine

DynState

A type-erased state of a state machine belonging to a module instance, see State

Executor

Executor that drives forward state machines under its management.

ExecutorBuilder

Builder to which module clients can be attached and used to build an Executor supporting these.

StateTransition

Represents one or multiple possible state transitions triggered in a common way

Traits

Context

Additional data made available to state machines of a module (e.g. API clients)

GlobalContext

Context given to all state machines

State

Implementors act as state machines that can be executed

Type Definitions

OperationId

Unique identifier for one semantic, correlatable operation.