Module fedimint_client::sm::executor

Structs

ActiveModuleOperationStateKeyPrefix 🔒

ActiveOperationStateKeyPrefix 🔒

ActiveStateKey

A state that is able to make progress eventually

ActiveStateKeyBytes

ActiveStateKeyPrefix

ActiveStateKeyPrefixBytes 🔒

ActiveStateMeta

Executor

Executor that drives forward state machines under its management.

ExecutorBuilder

Builder to which module clients can be attached and used to build an Executor supporting these.

ExecutorInner 🔒

InactiveModuleOperationStateKeyPrefix 🔒

InactiveOperationStateKeyPrefix 🔒

InactiveStateKey

A past or final state of a state machine

InactiveStateKeyBytes

InactiveStateKeyPrefix

InactiveStateKeyPrefixBytes 🔒

InactiveStateMeta

TransitionForActiveState 🔒

Enums

ActiveOrInactiveState 🔒

ExecutorDbPrefixes 🔒

Prefixes for executor DB entries

Constants

MAX_DB_ATTEMPTS 🔒

After how many attempts a DB transaction is aborted with an error

Type Aliases

ContextGen