Struct fedimint_client::Client

pub struct Client {
    pub(crate) inner: Arc<ClientInner>,
}

Fields
inner: Arc<ClientInner>
Implementations

impl Client

pub async fn create_backup(&self, metadata: Metadata) -> Result<ClientBackup>

Create a backup, include provided metadata

pub async fn create_encrypted_backup(
&self,
metadata: Metadata
) -> Result<EncryptedClientBackup>

Create a backup, include provided metadata, and encrypt it with a (derived) client root key

pub async fn backup_to_federation(&self, metadata: Metadata) -> Result<()>

Prepare an encrypted backup and send it to federation for storing

pub async fn wipe_state(&self) -> Result<()>

Wipe the client state (including module state)

pub async fn upload_backup(&self, backup: EncryptedClientBackup) -> Result<()>

Upload backup to federation

pub(crate) async fn restore_from_backup(&self) -> Result<Metadata>

Restore client state from backup download from federation (if found) or from scratch

This will restore (or initialize restoration process) in all sub-modules that support it.

pub async fn download_backup_from_federation(
&self
) -> Result<Option<ClientBackup>>

Download most recent valid backup found from the Federation

pub fn get_backup_id(&self) -> XOnlyPublicKey

Backup id derived from the root secret key (public key used to self-sign backup requests)

fn get_derived_backup_encryption_key_static(
secret: &DerivableSecret
) -> LessSafeKey

Static version of Self::get_derived_backup_encryption_key for testing without creating whole MintClient

fn get_derived_backup_signing_key_static(secret: &DerivableSecret) -> KeyPair

Static version of Self::get_derived_backup_signing_key for testing without creating whole MintClient

fn get_derived_backup_encryption_key(&self) -> LessSafeKey

fn get_derived_backup_signing_key(&self) -> KeyPair

pub async fn get_secret<S>(&self) -> S::Encodingwhere
S: RootSecretStrategy,

impl Client

pub fn builder() -> ClientBuilder

Initialize a client builder that can be configured to create a new client.

pub async fn start_executor(&self)

pub fn api(&self) -> &(dyn IGlobalFederationApi + 'static)

pub fn federation_id(&self) -> FederationId

pub fn get_internal_payment_markers(&self) -> Result<(PublicKey, u64)>

pub fn get_meta(&self, key: &str) -> Option<String>

pub fn decoders(&self) -> &ModuleDecoderRegistry

pub(crate) fn root_secret(&self) -> DerivableSecret

pub async fn finalize_and_submit_transaction<F, M>(
&self,
operation_id: OperationId,
operation_type: &str,
operation_meta: F,
tx_builder: TransactionBuilder
) -> Result<TransactionId>where
F: Fn(TransactionId, Option<OutPoint>) -> M + Clone + MaybeSend + MaybeSync,
M: Serialize + MaybeSend,

Add funding and/or change to the transaction builder as needed, finalize the transaction and submit it to the federation.

Errors

The function will return an error if the operation with given id already exists.

Panics

The function will panic if the the database transaction collides with other and fails with others too often, this should not happen except for excessively concurrent scenarios.

pub async fn add_state_machines(
&self,
dbtx: &mut DatabaseTransaction<'_>,
states: Vec<DynState<DynGlobalClientContext>>
) -> Result<()>

pub async fn get_active_operations(&self) -> HashSet<OperationId>

pub fn operation_log(&self) -> &OperationLog

pub fn get_first_module<M: ClientModule>(
&self,
module_kind: &ModuleKind
) -> (&M, ClientModuleInstance)

Returns a reference to a typed module client instance by kind

pub fn get_module_client_dyn(
&self,
instance_id: ModuleInstanceId
) -> Result<&(dyn IClientModule + Send + Sync)>

pub fn db(&self) -> &Database

pub async fn transaction_updates(
&self,
operation_id: OperationId
) -> TransactionUpdates

Returns a stream of transaction updates for the given operation id that can later be used to watch for a specific transaction being accepted.

pub fn get_first_instance(
&self,
module_kind: &ModuleKind
) -> Option<ModuleInstanceId>

Returns the instance id of the first module of the given kind. The primary module will always be returned before any other modules (which themselves are ordered by their instance id).

pub async fn root_secret_encoding<S>(&self) -> S::Encodingwhere
S: RootSecretStrategy,

Returns the data from which the client’s root secret is derived (e.g. BIP39 seed phrase struct).

pub async fn await_primary_module_output(
&self,
operation_id: OperationId,
out_point: OutPoint
) -> Result<Amount>

Waits for an output from the primary module to reach its final state.

pub fn get_config(&self) -> &ClientConfig

Returns the config with which the client was initialized.

pub fn primary_module(&self) -> &DynClientModule

Get the primary module

pub async fn get_balance(&self) -> Amount

Balance available to the client for spending

pub async fn subscribe_balance_changes(&self) -> BoxStream<'static, Amount>

Returns a stream that yields the current client balance every time it changes.

pub async fn discover_common_api_version(&self) -> Result<ApiVersionSet>

pub async fn discover_common_api_version_static(
config: &ClientConfig,
client_module_gen: &ClientModuleGenRegistry,
api: &DynGlobalApi
) -> Result<ApiVersionSet>

Query the federation for API version support and then calculate the best API version to use (supported by most guardians).

pub async fn supported_api_versions_summary_static(
config: &ClientConfig,
client_module_gen: &ClientModuleGenRegistry
) -> SupportedApiVersionsSummary

[SupportedApiVersionsSummary] that the client and its modules support

pub(crate) async fn load_and_refresh_common_api_version_static(
config: &ClientConfig,
module_gens: &ModuleGenRegistry<DynClientModuleGen>,
api: &DynGlobalApi,
db: &Database
) -> Result<ApiVersionSet>

Load the common api versions to use from cache and start a background process to refresh them.

This is a compromise so we not have to wait for version discovery to complete every time a Client is being built.

pub(crate) async fn refresh_common_api_version_static(
config: &ClientConfig,
module_gens: &ModuleGenRegistry<DynClientModuleGen>,
api: &DynGlobalApi,
db: &Database
) -> Result<ApiVersionSet>