Struct fedimint_client::Client

pub struct Client {
    pub(crate) inner: Arc<ClientInner>,
}

Fields
inner: Arc<ClientInner>
Implementations

impl Client

pub fn builder() -> ClientBuilder

Initialize a client builder that can be configured to create a new client.

pub fn api(&self) -> &(dyn IFederationApi + 'static)

pub async fn finalize_and_submit_transaction<F, M>(
&self,
operation_id: OperationId,
operation_type: &str,
operation_meta: F,
tx_builder: TransactionBuilder
) -> Result<TransactionId>where
F: Fn(TransactionId) -> M + Clone + MaybeSend + MaybeSync,
M: Serialize + MaybeSend,

Add funding and/or change to the transaction builder as needed, finalize the transaction and submit it to the federation.

pub async fn add_state_machines(
&self,
dbtx: &mut DatabaseTransaction<'_>,
states: Vec<DynState<DynGlobalClientContext>>
) -> Result<()>

pub async fn add_operation_log_entry(
&self,
dbtx: &mut DatabaseTransaction<'_>,
operation_id: OperationId,
operation_type: &str,
operation_meta: impl Serialize
)

pub async fn get_operations(
&self,
limit: usize
) -> Vec<(ChronologicalOperationLogKey, OperationLogEntry)>

Returns the last limit operations.

pub async fn get_operation(
&self,
operation_id: OperationId
) -> Option<OperationLogEntry>

pub fn get_module_client<M: ClientModule>(
&self,
instance_id: ModuleInstanceId
) -> Result<&M>

Returns a reference to a typed module client instance. Returns an error if the instance isn’t registered or the module kind doesn’t match.

pub fn db(&self) -> &Database

pub async fn transaction_updates(
&self,
operation_id: OperationId
) -> TransactionUpdates

Returns a stream of transaction updates for the given operation id that can later be used to watch for a specific transaction being accepted.

pub fn get_first_instance(
&self,
module_kind: &ModuleKind
) -> Option<ModuleInstanceId>

Returns the instance id of the first module of the given kind. The primary module will always be returned before any other modules (which themselves are ordered by their instance id).