Trait fedimint_client::IGlobalClientContext

pub trait IGlobalClientContext: Debug + MaybeSend + MaybeSync + 'static {
    fn module_api(&self) -> DynFederationApi;
    fn client_config(&self) -> &ClientConfig;
    fn api(&self) -> &DynFederationApi;
    fn decoders(&self) -> &ModuleDecoderRegistry;
    fn claim_input_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
        input: InstancelessDynClientInput
    ) -> Pin<Box<dyn Future<Output = TransactionId> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn fund_output_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
        output: InstancelessDynClientOutput
    ) -> Pin<Box<dyn Future<Output = Result<TransactionId>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn add_state_machine_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
        sm: Box<dyn IState<DynGlobalClientContext> + Send + Sync>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn transaction_update_stream<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = BoxStream<'_, OperationState<TxSubmissionStates>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Required Methods

fn module_api(&self) -> DynFederationApi

Returned a reference client’s module API client, so that module-specific calls can be made

fn client_config(&self) -> &ClientConfig

fn api(&self) -> &DynFederationApi

Returns a reference to the client’s federation API client. The provided interface IFederationApi typically does not provide the necessary functionality, for this extension traits like fedimint_core::api::GlobalFederationApi have to be used.

fn decoders(&self) -> &ModuleDecoderRegistry

fn claim_input_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
input: InstancelessDynClientInput
) -> Pin<Box<dyn Future<Output = TransactionId> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

This function is mostly meant for internal use, you are probably looking for DynGlobalClientContext::claim_input.

fn fund_output_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
output: InstancelessDynClientOutput
) -> Pin<Box<dyn Future<Output = Result<TransactionId>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

This function is mostly meant for internal use, you are probably looking for DynGlobalClientContext::fund_output.

fn add_state_machine_dyn<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ClientSMDatabaseTransaction<'life2, 'life3>,
sm: Box<dyn IState<DynGlobalClientContext> + Send + Sync>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Adds a state machine to the executor.

fn transaction_update_stream<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = BoxStream<'_, OperationState<TxSubmissionStates>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,