Trait fedimint_client::IGlobalClientContext

pub trait IGlobalClientContext: Debug + MaybeSend + MaybeSync + 'static {
    fn api(&self) -> &(dyn IFederationApi + 'static);
    fn finalize_and_submit_transaction<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        operation_id: OperationId,
        tx_builder: TransactionBuilder
    ) -> Pin<Box<dyn Future<Output = Result<TransactionId>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn await_active_state<'life0, 'async_trait>(
        &'life0 self,
        state: DynState<DynGlobalClientContext>
    ) -> Pin<Box<dyn Future<Output = ActiveState> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn await_inactive_state<'life0, 'async_trait>(
        &'life0 self,
        state: DynState<DynGlobalClientContext>
    ) -> Pin<Box<dyn Future<Output = InactiveState> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Required Methods

fn api(&self) -> &(dyn IFederationApi + 'static)

Returns a reference to the client’s federation API client. The provided interface IFederationApi typically does not provide the necessary functionality, for this extension traits like fedimint_core::api::GlobalFederationApi have to be used.

fn finalize_and_submit_transaction<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut DatabaseTransaction<'life2>,
operation_id: OperationId,
tx_builder: TransactionBuilder
) -> Pin<Box<dyn Future<Output = Result<TransactionId>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Add funding and/or change to the transaction builder as needed, finalize the transaction and submit it to the federation once dbtx is committed.

fn await_active_state<'life0, 'async_trait>(
&'life0 self,
state: DynState<DynGlobalClientContext>
) -> Pin<Box<dyn Future<Output = ActiveState> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Wait for any active state to appear in the database

fn await_inactive_state<'life0, 'async_trait>(
&'life0 self,
state: DynState<DynGlobalClientContext>
) -> Pin<Box<dyn Future<Output = InactiveState> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Wait for any inactive state to appear in the database